// frontend/src/shared/api/generated/models/PrepareLotteryFinalizationRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PrepareLotteryFinalizationRequest
 */
export interface PrepareLotteryFinalizationRequest {
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryFinalizationRequest
     */
    idempotencyKey: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryFinalizationRequest
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryFinalizationRequest
     */
    lotteryId: string;
}

/**
 * Check if a given object implements the PrepareLotteryFinalizationRequest interface.
 */
export function instanceOfPrepareLotteryFinalizationRequest(value: object): value is PrepareLotteryFinalizationRequest {
    if (!('idempotencyKey' in value) || value['idempotencyKey'] === undefined) return false;
    if (!('walletId' in value) || value['walletId'] === undefined) return false;
    if (!('lotteryId' in value) || value['lotteryId'] === undefined) return false;
    return true;
}

export function PrepareLotteryFinalizationRequestFromJSON(json: any): PrepareLotteryFinalizationRequest {
    return PrepareLotteryFinalizationRequestFromJSONTyped(json, false);
}

export function PrepareLotteryFinalizationRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareLotteryFinalizationRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'idempotencyKey': json['idempotencyKey'],
        'walletId': json['walletId'],
        'lotteryId': json['lotteryId'],
    };
}

export function PrepareLotteryFinalizationRequestToJSON(json: any): PrepareLotteryFinalizationRequest {
    return PrepareLotteryFinalizationRequestToJSONTyped(json, false);
}

export function PrepareLotteryFinalizationRequestToJSONTyped(value?: PrepareLotteryFinalizationRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'idempotencyKey': value['idempotencyKey'],
        'walletId': value['walletId'],
        'lotteryId': value['lotteryId'],
    };
}