// frontend/src/shared/api/generated/models/PrepareLotteryRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PrepareLotteryRequest
 */
export interface PrepareLotteryRequest {
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryRequest
     */
    idempotencyKey: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryRequest
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryRequest
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof PrepareLotteryRequest
     */
    ticketPrice: number;
    /**
     * 
     * @type {Date}
     * @memberof PrepareLotteryRequest
     */
    endAt: Date;
}

/**
 * Check if a given object implements the PrepareLotteryRequest interface.
 */
export function instanceOfPrepareLotteryRequest(value: object): value is PrepareLotteryRequest {
    if (!('idempotencyKey' in value) || value['idempotencyKey'] === undefined) return false;
    if (!('walletId' in value) || value['walletId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('ticketPrice' in value) || value['ticketPrice'] === undefined) return false;
    if (!('endAt' in value) || value['endAt'] === undefined) return false;
    return true;
}

export function PrepareLotteryRequestFromJSON(json: any): PrepareLotteryRequest {
    return PrepareLotteryRequestFromJSONTyped(json, false);
}

export function PrepareLotteryRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareLotteryRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'idempotencyKey': json['idempotencyKey'],
        'walletId': json['walletId'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'ticketPrice': json['ticketPrice'],
        'endAt': (new Date(json['endAt'])),
    };
}

export function PrepareLotteryRequestToJSON(json: any): PrepareLotteryRequest {
    return PrepareLotteryRequestToJSONTyped(json, false);
}

export function PrepareLotteryRequestToJSONTyped(value?: PrepareLotteryRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'idempotencyKey': value['idempotencyKey'],
        'walletId': value['walletId'],
        'name': value['name'],
        'description': value['description'],
        'ticketPrice': value['ticketPrice'],
        'endAt': value['endAt'].toISOString(),
    };
}