// frontend/src/shared/api/generated/models/PrepareLotteryTransaction.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PrepareLotteryTransactionTx } from './PrepareLotteryTransactionTx';
import {
    PrepareLotteryTransactionTxFromJSON,
    PrepareLotteryTransactionTxFromJSONTyped,
    PrepareLotteryTransactionTxToJSON,
    PrepareLotteryTransactionTxToJSONTyped,
} from './PrepareLotteryTransactionTx';
import type { TonNetwork } from './TonNetwork';
import {
    TonNetworkFromJSON,
    TonNetworkFromJSONTyped,
    TonNetworkToJSON,
    TonNetworkToJSONTyped,
} from './TonNetwork';

/**
 * 
 * @export
 * @interface PrepareLotteryTransaction
 */
export interface PrepareLotteryTransaction {
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryTransaction
     */
    intentId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareLotteryTransaction
     */
    lotteryId: string;
    /**
     * 
     * @type {TonNetwork}
     * @memberof PrepareLotteryTransaction
     */
    network: TonNetwork;
    /**
     * 
     * @type {Date}
     * @memberof PrepareLotteryTransaction
     */
    expiresAt: Date;
    /**
     * 
     * @type {PrepareLotteryTransactionTx}
     * @memberof PrepareLotteryTransaction
     */
    tx: PrepareLotteryTransactionTx;
}



/**
 * Check if a given object implements the PrepareLotteryTransaction interface.
 */
export function instanceOfPrepareLotteryTransaction(value: object): value is PrepareLotteryTransaction {
    if (!('intentId' in value) || value['intentId'] === undefined) return false;
    if (!('lotteryId' in value) || value['lotteryId'] === undefined) return false;
    if (!('network' in value) || value['network'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('tx' in value) || value['tx'] === undefined) return false;
    return true;
}

export function PrepareLotteryTransactionFromJSON(json: any): PrepareLotteryTransaction {
    return PrepareLotteryTransactionFromJSONTyped(json, false);
}

export function PrepareLotteryTransactionFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareLotteryTransaction {
    if (json == null) {
        return json;
    }
    return {
        
        'intentId': json['intentId'],
        'lotteryId': json['lotteryId'],
        'network': TonNetworkFromJSON(json['network']),
        'expiresAt': (new Date(json['expiresAt'])),
        'tx': PrepareLotteryTransactionTxFromJSON(json['tx']),
    };
}

export function PrepareLotteryTransactionToJSON(json: any): PrepareLotteryTransaction {
    return PrepareLotteryTransactionToJSONTyped(json, false);
}

export function PrepareLotteryTransactionToJSONTyped(value?: PrepareLotteryTransaction | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'intentId': value['intentId'],
        'lotteryId': value['lotteryId'],
        'network': TonNetworkToJSON(value['network']),
        'expiresAt': value['expiresAt'].toISOString(),
        'tx': PrepareLotteryTransactionTxToJSON(value['tx']),
    };
}